+++ src/controllers/recipeController.js
@@ -9,22 +9,26 @@ const PDFDocument = require('pdfkit');
 router.post('/create', async (req, res) => {
     const { title, category, imageUrl, ingredients, preparation, _ownerId } = req.body;
 
-    try {
-        if (title == "" || category == "" || imageUrl == "") {
-            throw "All fields are required!";
-        } else {
-            const publication = await recipeService.createRecipe({
-                title,
-                category,
-                imageUrl,
-                ingredients,
-                preparation,
-                _ownerId
-            });
-            res.json(publication);
+    if (req.headers['X-Authorization']) {
+        try {
+            if (title == "" || category == "" || imageUrl == "") {
+                throw "All fields are required!";
+            } else {
+                const publication = await recipeService.createRecipe({
+                    title,
+                    category,
+                    imageUrl,
+                    ingredients,
+                    preparation,
+                    _ownerId
+                });
+                res.json(publication);
+            }
+        } catch (err) {
+            res.json({ message: err });
         }
-    } catch (err) {
-        res.json({ message: err });
+    } else {
+        res.status(401).json('Unauthorized - You don\'t have permissions to do that!');
     }
 });
 
@@ -53,25 +57,37 @@ router.get('/author/:id', async (req, res) => {
 });
 
 router.put('/edit/:id', async (req, res) => {
-    const data = req.body;
-    const recipeId = req.params.id;
-    const editted = await recipeService.editRecipe(recipeId, data);
-    res.json(editted);
+    if (req.headers['X-Authorization']) {
+        const data = req.body;
+        const recipeId = req.params.id;
+        const editted = await recipeService.editRecipe(recipeId, data);
+        res.json(editted);
+    } else {
+        res.status(401).json('Unauthorized - You don\'t have permissions to do that!');
+    }
 });
 
 router.delete('/delete/:id', async (req, res) => {
-    const recipeId = req.params.id;
-    const deleted = await recipeService.deleteRecipe(recipeId);
-    res.json(deleted);
+    if (req.headers['X-Authorization']) {
+        const recipeId = req.params.id;
+        const deleted = await recipeService.deleteRecipe(recipeId);
+        res.json(deleted);
+    } else {
+        res.status(401).json('Unauthorized - You don\'t have permissions to do that!');
+    }
 });
 
 router.post('/save/:recipeId', async (req, res) => {
-    const { userId } = req.body;
-    const recipeId = req.params.recipeId;
-    const recipe = await recipeService.getOne(recipeId);
-    await authService.saveRecipe(userId, recipe);
-    const author = await authService.getUser(userId);
-    res.json(author.savedRecipes);
+    if (req.headers['X-Authorization']) {
+        const { userId } = req.body;
+        const recipeId = req.params.recipeId;
+        const recipe = await recipeService.getOne(recipeId);
+        await authService.saveRecipe(userId, recipe);
+        const author = await authService.getUser(userId);
+        res.json(author.savedRecipes);
+    } else {
+        res.status(401).json('Unauthorized - You don\'t have permissions to do that!');
+    }
 });
 
 router.get('/save/:userId', async (req, res) => {
@@ -81,15 +97,30 @@ router.get('/save/:userId', async (req, res) => {
 });
 
 router.post('/download', (req, res) => {
-    const data = req.body;
-    const doc = new PDFDocument();
-    const content = `${data.title}\n\nCategory: ${data.category}\n\nIngredients:\n${data.ingredients.join('\n')}\n\nPreparation:\n${data.preparation}`;
-    doc.title = 'Recipe PDF';
-    doc.text(content);
-    res.setHeader('Content-Disposition', `attachment; filename=${data.title}.pdf`);
-    res.setHeader('Content-Type', 'application/pdf');
-    doc.pipe(res);
-    doc.end();
+    if (req.headers['X-Authorization']) {
+        const data = req.body;
+        const doc = new PDFDocument();
+        const content = `${data.title}\n\nCategory: ${data.category}\n\nIngredients:\n${data.ingredients.join('\n')}\n\nPreparation:\n${data.preparation}`;
+        doc.title = 'Recipe PDF';
+        doc.text(content);
+        res.setHeader('Content-Disposition', `attachment; filename=${data.title}.pdf`);
+        res.setHeader('Content-Type', 'application/pdf');
+        doc.pipe(res);
+        doc.end();
+    } else {
+        res.status(401).json('Unauthorized - You don\'t have permissions to do that!');
+    }
+});
+
+router.post('/comments', async (req, res) => {
+    const newComment = await recipeService.createComment(req.body);
+    res.json(newComment);
+});
+
+router.get('/comments/:recipeId', async (req, res) => {
+    const recipeId = req.params.recipeId;
+    const comments = await recipeService.getAllComments(recipeId);
+    res.json(comments);
 });
 
 module.exports = router;
